Don't try to guess whether to use GET or POST
Testing this via [ -t ] does not fully work since subshells inherit the
file descriptor target. There may still be a way to detect this (perhaps
by reading a byte from stdin to see if it's empty or not). I'll revisit
this later if I still care. :-)

Thanks to lhunath in #bash for the help.