Fix dash-isms
Fixes #16.

Dash does not like multiple assignments after a `local` keyword; each
must be a separate assignment otherwise the subsequent vars will be
globals.

Dash does not preserve whitespace or odd characters for direct
variable assignments in the context of the `local` keyword. I.e.,
`varname=$1` is fine but `local varname=$1` is not.