Match headers on crlf instead of cr
This matches the spec. Although it did work previously, I wonder if this
may solve some issues reported where the headers and body weren't
getting parsed apart correctly.

The spec also allows for a single line feed separator as well so we may
want to add a check for that in the future.